Remove package .namedays and move everything to .events.namedays
