+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/GreekSpecialNamedaysStrategy.java
@@ -1,7 +1,6 @@
 package com.alexstyl.specialdates.events.namedays;
 
 import com.alexstyl.specialdates.events.DayDate;
-import com.alexstyl.specialdates.namedays.NameCelebrations;
 
 import java.util.ArrayList;
