+++ mobile/src/main/java/com/alexstyl/specialdates/events/namedays/NamedayCalendar.java
@@ -1,8 +1,6 @@
 package com.alexstyl.specialdates.events.namedays;
 
 import com.alexstyl.specialdates.events.DayDate;
-import com.alexstyl.specialdates.namedays.Dates;
-import com.alexstyl.specialdates.namedays.NameCelebrations;
 
 import java.util.ArrayList;
 import java.util.List;
