+++ mobile/src/main/java/com/alexstyl/specialdates/search/NamedayCard.java
@@ -1,7 +1,7 @@
 package com.alexstyl.specialdates.search;
 
 import com.alexstyl.specialdates.events.DayDate;
-import com.alexstyl.specialdates.namedays.NameCelebrations;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
 
 public class NamedayCard {
