+++ mobile/src/main/java/com/alexstyl/specialdates/search/NamedaysLoader.java
@@ -7,7 +7,7 @@ import com.alexstyl.specialdates.events.namedays.NamedayCalendar;
 import com.alexstyl.specialdates.events.namedays.NamedayCalendarProvider;
 import com.alexstyl.specialdates.events.namedays.NamedayLocale;
 import com.alexstyl.specialdates.events.namedays.NamedayPreferences;
-import com.alexstyl.specialdates.namedays.NameCelebrations;
+import com.alexstyl.specialdates.events.namedays.NameCelebrations;
 import com.alexstyl.specialdates.ui.loader.SimpleAsyncTaskLoader;
 
 public class NamedaysLoader extends SimpleAsyncTaskLoader<NameCelebrations> {
