+++ mobile/src/test/java/com/alexstyl/specialdates/namedays/CharacterNodeTest.java
@@ -1,5 +1,8 @@
 package com.alexstyl.specialdates.namedays;
 
+import com.alexstyl.specialdates.events.namedays.CharacterNode;
+import com.alexstyl.specialdates.events.namedays.Node;
+
 public class CharacterNodeTest extends NodeTest {
 
     @Override
