+++ mobile/src/test/java/com/alexstyl/specialdates/namedays/SoundNodeTest.java
@@ -1,5 +1,8 @@
 package com.alexstyl.specialdates.namedays;
 
+import com.alexstyl.specialdates.events.namedays.Node;
+import com.alexstyl.specialdates.events.namedays.SoundNode;
+
 public class SoundNodeTest extends NodeTest {
 
     @Override
